+++ lib/src/code_generators/swagger_additions_generator.dart
@@ -90,6 +90,9 @@ ${options.overrideToString ? "import 'dart:convert';" : ''}
     if (enumsImport.isNotEmpty) {
       result.write(enumsImport);
     }
+    for (var element in options.importPaths) {
+      result.write('import \'$element\';');
+    }
 
     if (enumsExport.isNotEmpty) {
       result.write(enumsExport);