+++ lib/src/models/generator_options.g2.dart
@@ -44,7 +44,12 @@ GeneratorOptions _$GeneratorOptionsFromJson(Map json) => GeneratorOptions(
           json['use_path_for_request_names'] as bool? ?? true,
       includeIfNull: json['include_if_null'] as bool?,
       modelPostfix: json['model_postfix'] as String? ?? '',
+      customReturnType: json['custom_return_type'] as String? ?? '',
       includePaths: (json['include_paths'] as List<dynamic>?)
+              ?.map((e) => e as String)
+              .toList() ??
+          [],
+      importPaths: (json['import_paths'] as List<dynamic>?)
               ?.map((e) => e as String)
               .toList() ??
           [],
@@ -111,6 +116,8 @@ Map<String, dynamic> _$GeneratorOptionsToJson(GeneratorOptions instance) =>
       'default_header_values_map': instance.defaultHeaderValuesMap,
       'response_override_value_map': instance.responseOverrideValueMap,
       'include_paths': instance.includePaths,
+      'import_paths': instance.importPaths,
+      'custom_return_type': instance.customReturnType,
       'exclude_paths': instance.excludePaths,
     };