+++ lib/src/swagger_code_generator.dart
@@ -60,30 +60,6 @@ class SwaggerCodeGenerator {
       _getSwaggerEnumsGenerator(root, options)
           .generateAllEnums(fileName: fileName, root: root);
 
-  // String generateResponses(
-  //   SwaggerRoot root,
-  //   String fileName,
-  //   GeneratorOptions options,
-  //   List<EnumModel> allEnums,
-  // ) =>
-  //     _getSwaggerModelsGenerator(root, options).generateResponses(
-  //       root: root,
-  //       fileName: fileName,
-  //       allEnums: allEnums,
-  //     );
-
-  String generateRequestBodies(
-    SwaggerRoot root,
-    String fileName,
-    GeneratorOptions options,
-    List<EnumModel> allEnums,
-  ) =>
-      _getSwaggerModelsGenerator(root, options).generateRequestBodies(
-        root: root,
-        fileName: fileName,
-        allEnums: allEnums,
-      );
-
   String generateEnums(
     SwaggerRoot root,
     String fileName,