+++ lib/swagger_dart_code_generator.dart
@@ -217,13 +217,6 @@ class SwaggerDartCodeGenerator implements Builder {
       allEnums,
     );
 
-    final requestBodies = codeGenerator.generateRequestBodies(
-      contents,
-      fileWithoutExtension,
-      options,
-      allEnums,
-    );
-
     final enums = codeGenerator.generateEnums(
       contents,
       fileWithoutExtension,
@@ -233,7 +226,7 @@ class SwaggerDartCodeGenerator implements Builder {
 
     final imports = codeGenerator.generateImportsContent(
       fileNameWithoutExtension,
-      models.isNotEmpty || requestBodies.isNotEmpty,
+      models.isNotEmpty,
       options.buildOnlyModels,
       enums.isNotEmpty,
       options.separateModels,
@@ -265,7 +258,6 @@ class SwaggerDartCodeGenerator implements Builder {
               imports,
               requests,
               options.separateModels ? '' : models,
-              options.separateModels ? '' : requestBodies,
               customDecoder,
               options.separateModels ? '' : dateToJson));
     }
@@ -286,7 +278,6 @@ class SwaggerDartCodeGenerator implements Builder {
       ///Write models to separate file
       final formattedModels = _tryFormatCode(_generateSeparateModelsFileContent(
         models,
-        requestBodies,
         fileNameWithoutExtension,
         dateToJson,
         enums.isNotEmpty,
@@ -305,7 +296,6 @@ class SwaggerDartCodeGenerator implements Builder {
       String imports,
       String requests,
       String models,
-      String requestBodies,
       String customDecoder,
       String dateToJson) {
     final result = """
@@ -314,8 +304,6 @@ $imports
 ${options.buildOnlyModels ? '' : requests}
 
 $models
-
-$requestBodies
 
 ${options.buildOnlyModels ? '' : customDecoder}
 
@@ -363,7 +351,6 @@ $dateToJson
 
   String _generateSeparateModelsFileContent(
     String models,
-    String requestBodies,
     String fileNameWithoutExtension,
     String dateToJson,
     bool hasEnums,
@@ -389,8 +376,6 @@ $overridenModels
     part '$fileNameWithoutExtension.models.swagger.g.dart';
 
     $models
-
-    $requestBodies
 
     $dateToJson
     ''';